+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### Fixed
+
+- Fix Toggle component with `cls` option in hover mode
+- Fix `aria-roledescription` attribute in Slider nav component
+
 ## 3.16.19 (May 24, 2023)
 
 ### Fixed